Document how to serialize nodes to protobuf.
Fixes #99. Also included in this commit are a couple of minor typo fixes.  Note
that the example creates a slice for which there isn't a protobuf analogue; a
better solution would be to generate a wrapper with a repeated field, but this
at least illustrates the serialization API.

This replaces #105 that I apparently closed without merging it >.<

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>